EdgeBundling fixes and optimizations: use anonymous ephemeral properties instead of named ones + use IntegerProperty instead of DoubleProperty when only integer values are needed
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@11895 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/general/EdgeBundling/Dijkstra.cpp
@@ -152,7 +152,7 @@ void Dijkstra::initDijkstra(const tlp::Graph *const forbidden, tlp::node srcTlp,
   resultEdges.setAll(false);
 }
 //=======================================================================
-void Dijkstra::searchPaths(node ntlp, DoubleProperty *depth) {
+void Dijkstra::searchPaths(node ntlp, IntegerProperty *depth) {
 
   node n = ntlp2dik.get(ntlp);
 
@@ -178,7 +178,7 @@ void Dijkstra::searchPaths(node ntlp, DoubleProperty *depth) {
       continue;
 
     resultEdges[e] = true;
-    double dep = depth->getEdgeValue(edik2tlp[e]) + 1.;
+    int dep = depth->getEdgeValue(edik2tlp[e]) + 1;
 #ifdef _OPENMP
 #pragma omp critical(DEPTH)
 #endif
